Updated ride page for new decisions on when to display images, changed ScheduleBar to be used on all ride pages
