+++ lib/pages/Home.dart
@@ -3,7 +3,7 @@ import 'package:carriage_rider/pages/Request_Ride_Loc.dart';
 import 'package:carriage_rider/providers/AuthProvider.dart';
 import 'package:carriage_rider/pages/Notifications.dart';
 import 'package:carriage_rider/pages/Profile.dart';
-import 'package:carriage_rider/pages/Upcoming.dart';
+import 'package:carriage_rider/pages/RidePage.dart';
 import 'package:carriage_rider/utils/app_config.dart';
 import 'package:carriage_rider/providers/RidesProvider.dart';
 import 'package:flutter/material.dart';
@@ -13,6 +13,8 @@ import 'package:carriage_rider/pages/Settings.dart';
 import 'package:carriage_rider/pages/Contact.dart';
 import '../models/RideObject.dart';
 import 'package:carriage_rider/utils/CarriageTheme.dart';
+
+import 'Upcoming.dart';
 
 void main() {
   MaterialApp(routes: {
