+++ lib/pages/Settings.dart
@@ -1,7 +1,7 @@
 import 'dart:ui';
 import 'package:carriage_rider/providers/AuthProvider.dart';
 import 'package:carriage_rider/models/Location.dart';
-import 'package:carriage_rider/pages/Upcoming.dart';
+import 'package:carriage_rider/pages/RidePage.dart';
 import 'package:carriage_rider/utils/CarriageTheme.dart';
 import 'package:carriage_rider/utils/app_config.dart';
 import 'package:flutter/cupertino.dart';
