+++ lib/widgets/ScheduleBar.dart
@@ -1,39 +1,25 @@
 import 'package:flutter/material.dart';
 
-import '../pages/Home.dart';
-
 class ScheduleBar extends StatelessWidget implements PreferredSizeWidget {
+
+  ScheduleBar(this.textColor, this.backgroundColor);
+  final Color textColor;
+  final Color backgroundColor;
+
   static final double prefHeight = 80;
   @override
   Widget build(BuildContext context) {
-    return Container(
-      height: prefHeight,
-      child: Column(
-        children: [
-          Expanded(child: SizedBox()),
-          GestureDetector(
-            onTap: () => Navigator.of(context).pushReplacement(
-                MaterialPageRoute(builder: (BuildContext context) => Home())
-            ),
-            child: Row(
-                crossAxisAlignment: CrossAxisAlignment.center,
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: [
-                  GestureDetector(
-                    child: Padding(
-                      padding: EdgeInsets.only(left: 16, right: 8),
-                      child: Icon(Icons.arrow_back_ios),
-                    ),
-                    onTap: () {
-                      Navigator.of(context).pushReplacement(
-                          MaterialPageRoute(builder: (BuildContext context) => Home())
-                      );
-                    },
-                  ),
-                  Text('Schedule', style: TextStyle(fontSize: 17))
-                ]),
-          ),
-        ],
+    return AppBar(
+      title: Text(
+        'Schedule',
+        style: TextStyle(
+            color: textColor, fontSize: 20, fontFamily: 'SFPro'),
+      ),
+      backgroundColor: backgroundColor,
+      titleSpacing: 0.0,
+      leading: IconButton(
+        icon: Icon(Icons.arrow_back_ios, color: textColor),
+        onPressed: () => Navigator.pop(context, false),
       ),
     );
   }
